+++ src/test/project/src/renametests/util/ExtendFormatter.js
@@ -1,0 +1,10 @@
+sap.ui.define([
+	"com/test/renametests/util/Formatter"
+], function(
+	Formatter
+) {
+	"use strict";
+
+	return Formatter.extend("com.test.renametest.util.ExtendFormatter", {
+	});
+});